- [PROT] Pykeops

+++ gempy_engine/core/data/options.py
@@ -9,8 +9,8 @@ class InterpolationOptions:
     range: int
     c_o: float
     uni_degree: int = 1
-    i_res: float = 4.
-    gi_res: float = 2.
+    i_res: int = 4
+    gi_res: int = 2
     number_dimensions: int = 3
     number_octree_levels:int = 1
     kernel_function: AvailableKernelFunctions = AvailableKernelFunctions.exponential
